show the list of previous days but wrong day

+++ app/src/main/java/com/iut/meteoreo/data/DayTemperature.kt
@@ -2,6 +2,6 @@ package com.iut.meteoreo.data
 
 data class DayTemperature(
     val day: String,
-    val maxTemperature: Int,
-    val minTemperature: Int
+    val maxTemperature: Double,
+    val minTemperature: Double
 )
